wirefrome done

+++ src/Usbdevice.c
@@ -285,7 +285,7 @@ void printBuffer(char* buffer){
     printf("buffer = [");
     char i;
     for(i=0;i<BUFLEN_SERVO_DATA;i++){
-        printf("%d, ", buffer[i]);
+        printf("%d, ", (uint8_t)buffer[i]);
     }
     printf("]\n");
 }
